pyproject.toml
[project]
name = "witty-caption-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "motor>=3.7.1",
    "openai>=2.32.0",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "uvicorn>=0.44.0",
]
